Move all Opt files inside kfusion
