+++ kfusion/include/opt/resource.h
