+++ kfusion/src/mLibSource.cpp
@@ -1,5 +1,4 @@
-
-#include "mLibInclude.h"
+#include "opt/mLibInclude.h"
 
 #include "mLibCore.cpp"
 #include "mLibLodePNG.cpp"
